# frontend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Turk platform
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: submissions
    description: Screenshot submission operations
  - name: dashboard
    description: Dashboard summary operations
  - name: Storage
    description: Object storage upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List user submissions
      description: Returns all submissions for the authenticated user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [in_review, processed, accepted]
          description: Filter by status
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Page number for pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Create a new submission
      description: Creates a new screenshot submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      description: Returns aggregated stats for the authenticated user's dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /dashboard/recent:
    get:
      operationId: getRecentSubmissions
      tags: [dashboard]
      summary: Get recent submissions for dashboard
      description: Returns the 5 most recent submissions for the authenticated user
      responses:
        "200":
          description: Recent submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Submission:
      type: object
      required: [id, userId, imageUrl, objectPath, status, points, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        imageUrl:
          type: string
        objectPath:
          type: string
        fileName:
          type: ["string", "null"]
        platform:
          type: ["string", "null"]
          description: Social media platform (e.g. Instagram, Twitter, TikTok)
        status:
          type: string
          enum: [in_review, processed, accepted]
        points:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    SubmissionInput:
      type: object
      required: [imageUrl, objectPath]
      properties:
        imageUrl:
          type: string
          description: Public URL of the uploaded image
        objectPath:
          type: string
          description: GCS object path returned from presigned URL request
        fileName:
          type: string
          description: Original file name
        platform:
          type: string
          description: Social media platform (e.g. Instagram, Twitter, TikTok)

    SubmissionList:
      type: object
      required: [submissions, total, page, limit]
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    DashboardSummary:
      type: object
      required: [totalPoints, totalSubmissions, accepted, inReview, processed, updatedToday]
      properties:
        totalPoints:
          type: integer
        totalSubmissions:
          type: integer
        accepted:
          type: integer
        inReview:
          type: integer
        processed:
          type: integer
        updatedToday:
          type: integer

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string